update access to calendar component
