+++ packages/Calendar/src/index.js
@@ -1,0 +1,92 @@
+// @flow
+import React, { useState, useCallback } from 'react';
+import { View, StyleSheet, Dimensions } from 'react-native';
+import Header from './Header';
+import Month from './Month';
+
+const styles = StyleSheet.create({
+  container: {
+    flexDirection: 'row',
+    backgroundColor: 'white',
+  },
+});
+
+function getMonthCount({ width }, date) {
+  return width > 400
+    ? [date, new Date(date.getFullYear(), date.getMonth() + 1)] : [date];
+}
+
+type Props = {
+  renderDate: (date: Date) => React.Node,
+  setValue: number => void,
+  value: ?Date,
+}
+
+function Calendar({
+  renderDate, setValue, value,
+}: Props) {
+  const [months, setMonths] = useState(() => getMonthCount(Dimensions.get('screen'), value));
+
+  const handleLayout = useCallback((e) => {
+    setMonths(getMonthCount(e.nativeEvent.layout, value));
+  }, [value, setMonths]);
+
+  const prevMonth = useCallback(() => {
+    setMonths(mnths => mnths.map(d => new Date(d.getFullYear(), d.getMonth() - 1, 1)));
+  }, [setMonths]);
+
+  const nextMonth = useCallback(() => {
+    setMonths(mnths => mnths.map(d => new Date(d.getFullYear(), d.getMonth() + 1, 1)));
+  }, [setMonths]);
+
+  const selectDate = useCallback((d, long) => {
+    setValue((prev) => {
+      if (Array.isArray(prev)) {
+        // logic to remove the date if already Selected
+        const isAlreadySelected = prev.filter(dates => dates.toDateString() === d.toDateString());
+        return isAlreadySelected.length > 0
+          ? [...prev.filter(dates => dates.toDateString() !== d.toDateString())]
+          : [...prev, d];
+      }
+      if (long) {
+        return [prev, d];
+      }
+      return d;
+    });
+  }, [setValue]);
+
+  const removeMultipleSelect = useCallback(() => {
+    setValue((prev) => {
+      return prev.pop();
+    });
+  }, [setValue]);
+
+
+  const first = 0;
+  const last = months.length - 1;
+
+  return (
+    <View style={styles.container} onLayout={handleLayout}>
+      {months.map((month, idx) => (
+        <View key={month} style={{ flex: 1, padding: 5 }}>
+          <Header
+            date={month}
+            prevMonth={(idx === first || months.length === 1) && prevMonth}
+            nextMonth={(idx === last || months.length === 1) && nextMonth}
+            multiple={Array.isArray(value)}
+            removeMultipleSelect={removeMultipleSelect}
+          />
+          <Month
+            date={month}
+            value={value}
+            renderDate={renderDate}
+            selectDate={selectDate}
+          />
+        </View>
+      ))}
+    </View>
+  );
+}
+
+
+export default Calendar;
