+++ showcase/screens/Calendar.js
@@ -1,6 +1,6 @@
 import React, { useState } from 'react';
 import { View } from 'react-native';
-import CalendarWidget from '../../Calendar';
+import CalendarWidget from '@sharingapples/calendar';
 
 export default function Calendar() {
   const [value, setValue] = useState(new Date());
